making helper file
