+++ lib/extend-context.js
@@ -1,31 +1,7 @@
-import { createSigner, createVerifier } from 'fast-jwt';
 import _ from 'lodash';
 import Entity from './entity.js';
 import Collection from './collection.js';
-
-const JWT_SECRET = process.env.JWT_SECRET;
-const PRIMARY_JWT_OPTIONS = {
-  key: JWT_SECRET,
-  algorithm: process.env.JWT_ALGO || 'ES256',
-  iss: process.env.JWT_ISSUER
-};
-if (process.env.JWT_AUDIENCE) {
-  PRIMARY_JWT_OPTIONS.audience = process.env.JWT_AUDIENCE;
-}
-const signerSync = createSigner(PRIMARY_JWT_OPTIONS);
-const primaryVerifier = createVerifier({
-  key: JWT_SECRET,
-  algorithms: PRIMARY_JWT_OPTIONS.algorithm,
-  allowedIss: PRIMARY_JWT_OPTIONS.iss
-});
-let secondaryVerifier;
-if (process.env.JWT_SECONDARY_SECRET) {
-  secondaryVerifier = createVerifier({
-    key: process.env.JWT_SECONDARY_SECRET,
-    algorithms: process.env.JWT_SECONDARY_ALGO,
-    allowedIss: process.env.JWT_SECONDARY_ISSUER
-  });
-}
+import { issueToken, verifyToken } from './jwt-helpers.js';
 
 const toBool = function(value) {
   return !!(value === true || value === 'true' || value === '1' || value === 'yes');
@@ -145,27 +121,11 @@ export default function(event, context) {
   };
 
   // convenience method to verify a JWT token
-  context.verifyToken = function(token) {
-    let decodedToken, err;
-    try {
-      decodedToken = primaryVerifier(token);
-    } catch (e) {
-      err = e;
-    }
-    if (!decodedToken && secondaryVerifier) {
-      try {
-        decodedToken = secondaryVerifier(token);
-      } catch {}
-    }
-    if (err) {
-      throw err;
-    }
-    return decodedToken;
-  };
+  context.verifyToken = verifyToken;
 
   // convenience method to issue a JWT token
-  context.issueToken = function(payload) {
-    return signerSync(payload);
+  context.issueToken = (payload) => {
+    return issueToken(payload);
   };
 
   // convenience method for checking the current user's authorization
